app/requirements.txt
Flask
mysql-connector
setuptools
apscheduler
pytest
python-dotenv